+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+Dev
+# BS 03.12.2014
+* update Event Handling CAM-1422 
+** add "contact:update" in FE
+** add "friendRequest:rejected" in FE
+** add "identity:new" in FE
+
 0.4.4.
+# BS 03.12.2014
+* fix CAM-1478, last-message:read will only send, if message is decrypted
 # BS 28.11.2014
 * add Authentication Route to Desktop Modus, CAM-1374